notification

+++ frontend/eduwave-ui/src/components/navbar/AdminNavbar.jsx
@@ -3,13 +3,24 @@ import { Link, useNavigate } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSignOutAlt } from "@fortawesome/free-solid-svg-icons";
 import "./Navbar.css";
+import swal from "sweetalert"; 
 
 function Navbar() {
   const navigate = useNavigate();
   const handleClick = () => {
-    localStorage.clear();
-    navigate("/");
-    console.log("logout");
+    swal({
+      title: "Are you sure?",
+      text: "You will be logged out!",
+      icon: "warning",
+      buttons: ["Cancel", "Logout"],
+      dangerMode: true,
+    }).then((willLogout) => {
+      if (willLogout) {
+        localStorage.clear();
+        navigate("/");
+        console.log("logout");
+      }
+    });
   };
 
   return (
