+++ frontend/eduwave-ui/src/components/navbar/Navbar.jsx
@@ -3,15 +3,25 @@ import { Link, useNavigate } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSignOutAlt } from "@fortawesome/free-solid-svg-icons";
 import "./Navbar.css";
+import swal from "sweetalert"; 
 
 function Navbar() {
   const navigate = useNavigate();
   const handleClick = () => {
-    localStorage.clear();
-    navigate("/");
-    console.log("logout");
+    swal({
+      title: "Are you sure?",
+      text: "You will be logged out!",
+      icon: "warning",
+      buttons: ["Cancel", "Logout"],
+      dangerMode: true,
+    }).then((willLogout) => {
+      if (willLogout) {
+        localStorage.clear();
+        navigate("/");
+        console.log("logout");
+      }
+    });
   };
-
   return (
     <div className="navbar-container">
       <nav className="navbar">
@@ -25,7 +35,7 @@ function Navbar() {
           ></link>
           <li className="nav-item">
             <Link to="/enroll" className="navtext">
-              ENROLL
+              COURSES
             </Link>
           </li>
           <li className="nav-item">
